moved api tester logic to a separate class
